Drop support for engines-provided signature verification.
There are some checks that are done in |RSA_verify|, some checks done
in |RSA_verify_raw|, and some more checks done in
|RSA_default_verify_raw|. These are all part of the call chain when
verifying a certificate when an RSA engine isn't in use, and so all
their checks are done in the default case.

If an RSA engine that provides a |verify| method *were* in use, should
any of the checks being done in |RSA_verify_raw| be done? If an RSA
engine that provides a |verify_raw| method *were* in use, should any of
the checks being done in |rsa_default_verify_raw| be done? Are any of
the checks in these functions redundant or conflicting?

Rather than answer that first question, and to prepare for simplifying
the code in the future to address the second question, drop support for
engine-provided signature verification. It isn't necessary because,
unlike the private key operations, we always have the key material
needed to do the computation. Further, there are no side-channel
concerns or other issues other than, potentially, speed, where an
engine can do a better job. And, in general, we just do not need this
support.

Do the same, for the same reasons, for ECDSA signature verification.

Change-Id: Iaad8369d403082b728c831167cc386fdcabfb067