Update patch set 2
Patch Set 2:

David, the thing I'm unsure about, still, is whether all we *actually* have the public key material in the RSA struct in all cases where engines are in use? It seems like for this kind of change to be made, the meaning of the "OPAQUE" flag would also need to be changed?

Patch-set: 2
Label: Code-Review=0