Update patch set 3
Patch Set 3:

> Piotr, usually such proxing is needed for private key operations
 > and/or maybe encryption. Usually such proxing isn't too useful for
 > verification.
 
Agreed, but it seems weird to limit ENGINEs to subset of crypto ops, especially when the argument for doing so isn't very strong.

Note that this is my personal opinion and not BoringSSL's ;)

 > However, also keep in mind that I wrote this patch in full
 > ignorance of what Google uses BoringSSL for, except that I checked
 > that it wouldn't hurt the Android stuff, AFAICT. I would love it if
 > you could share links to other uses of this mechanism.

Quite frankly, I don't know too much about the internal use cases... I based my comment mostly on my knowledge about existing open-source software that uses ENGINEs for fancy stuff, i.e. OpenBSD is using it for SSL privsep, so that unprivileged processes don't have access to the private key (but you are right, they use it only for encryption/decryption... and yes, they are using LibreSSL, not BoringSSL).

Patch-set: 3