Update patch set 3
Patch Set 3:

Google's stuff (AFAICT) does not use the engine stuff for all the same reasons as OpenSSL does. Thus, it makes sense for for it to be simpler, since BoringSSL is supposed to be optimized for what Google needs.

Note that this is just the first of multiple changes I will (probably) submit to reduce the engine mechanism down to the minimum needed and to simplify the RSA code.

Patch-set: 3
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0