Make bssl_shim build against OpenSSL instead.
Not for landing, but uploaded so people can play with it. For the most
part this is removing BoringSSL-specific features and handling a few API
divergences.

(The DH_get_2048_256 API difference seems to have been an mistake on our
part when importing. I'll fix that one.)

To build, cd into ssl/test and run:

  g++ -std=c++11 -I$openssl_dir/include *.cc $openssl_dir/libcrypto.a $openssl_dir/libssl.a -o ossl_shim

Then cd into ssl/test/runner

  go test -shim-path ../ossl_shim

Change-Id: I74d832207b970f1910b1baf6abd20e5f255aecee