Update patch set 3
Patch Set 3:

> Google's stuff (AFAICT) does not use the engine stuff for all the
 > same reasons as OpenSSL does. Thus, it makes sense for for it to be
 > simpler, since BoringSSL is supposed to be optimized for what
 > Google needs.
 > 
 > Note that this is just the first of multiple changes I will
 > (probably) submit to reduce the engine mechanism down to the
 > minimum needed and to simplify the RSA code.

It'd be good to give me a heads-up ahead of time if you're planning these kinds of changes, by the way. If nothing else, so I have some context as to what you're up to. And I might be able to tell you ahead of time if there'll be some problem somewhere. (I'd usually chat with Adam briefly before embarking on a large change to make sure we were on the same page.)

(Still haven't looked at the CL. Will do that next week. I hate doing code reviews and won't do them when I'm not paid to. :-) )

Patch-set: 3