SSL_set_fd should create socket BIOs, not fd BIOs.
In OpenSSL, they create socket BIOs. The distinction isn't really important on
UNIX, but it is on Windows.

TODO update documentation and such. Unfortunately, this also proliferates the
int vs SOCKET confusion in the API. Enough software assumes one can stuff
SOCKETs into ints that I doubt Microsoft will ever break compatibility there,
but it'd still be good to make it possible to avoid that assumption.

Change-Id: I67408c218572228cb1a7d269892513cda4261c82