Update patch set 3
Patch Set 3:

> > This is a lot of rather ugly typedefs to work around a compiler
 > bug for what seems to be a one-off replacement for probably the
 > least objectionable bit of perlasm in the project.
 > 
 > I agree 100%. Again, I only want to solve this problem once, not
 > every time it occurs. Ugliness is ugly. A type cast is only a few
 > characters, so aesthetically it would be more pleasing, but I think
 > that would be choosing style over substance. (Again, my goal is to
 > remove all the type casts in the code base, as far as is
 > practical.)

What is "this problem"? As an abstraction, uint64_alt isn't meaningful. When do I use uint64_alt? When do I use uint64_t? What if I have a platform where uint64_t is defined as unsigned long long, but unsigned long is also the right size?

In contrast, uint64_t is meaningful. "uint64_t a 64-bit integer so you don't have to care about long vs long long." One can't actually describe uint64_alt as a meaningful thing.

It seems to me the problem is simply that the particular intrinsic you want to use (_rdrand64_step) is defined to take a unsigned long long, right? In that case, that is clearly a very local issue. In fact, if it's defined in that way, you should just use unsigned long long and sizeof(unsigned long long) and avoid this mess of typedefs. Or maybe if you want to abstract between _rdrand32_step and _rdrand64_step, in the same place, abstract their return values into a rdrand_return_t or something typedef and sizeof that.

The point is this is a local problem. It needs a local solution.

Patch-set: 3