Update patch set 3
Patch Set 3:

> There are some places in the code that assume |size_t| is what I
 > defined native_uint to be. One example is https://github.com/briansmith/ring/blob/boringssl/crypto/modes/internal.h#L168
 > and the related code in gcm.c, and the similar code throughout
 > crypto/modes and probably other places that I've forgotten.

That's not actually using those #ifdefs, is it? I suppose it would end up using a suboptimal size on x32, though it does still work. *shrug* I see no other module that cares about sizeof(size_t) in this way. It seems to entirely come from crypto/modes want of playing games with alignment.

Regardless, my main objection is with adding this NATIVE_WORD_SIZE redundant to OPENSSL_{32,64}_BIT. I suspect as things stand, x86-64 NaCl looks like OPENSSL_64_BIT but pointers happen to smaller. Given it's just this one module which is affected (and not in correctness), this seems the best option, and so NATIVE_WORD_SIZE is purely redundant.

(That said, I'm also not a fan of the complication of having yet another global abstraction to worry about.)

Patch-set: 3