Update patch set 2
Patch Set 2: Code-Review+2

If I understand correctly, the issue is |r| is left with uninitialized memory on some failing returns, right? So any code which was sensitive to that is likely still broken because OpenSSL code tends to leave outputs in an undefined state on error (just still freeable).

(This version is clearly better, of course. I just wanted to make sure I'm understanding the impact correctly.)

Patch-set: 2
Label: Code-Review=+2