Explicitly check SSL_SENT_SHUTDOWN at entry points.
There is a SSL_SENT_SHUTDOWN deep in ssl3_read_bytes. It is necessary
for now because that's the only way ssl3_read_bytes knows it is called
from ssl3_read_close_notify. But it also, as a side effect, changes the
behavior of SSL_read, etc., when called from SSL_shutdown.

In preparation for splitting up ssl3_read_bytes, remove all other entry
points reliance on this check. Instead, it is an error to call SSL_read,
etc., after SSL_shutdown since that would be silly.

The goal is to move all the record-type-mismatch logic up into
read_{app_data,change_cipher_spec,close_notify} and get_message as it is almost
entirely context-specific. The only common case is alert processing, which can
be factored out. At that point, read_bytes can finally fall away.

Change-Id: Ie7a7e5520b464360b1e6316c34ec9854b571782f