Consolidate SSL_RECEIVED_SHUTDOWN checks.
SSL_RECEIVED_SHUTDOWN checks in the record layer happen in two different
places. Some operations (but not all) check it, and so does read_bytes. Move it
to get_record.

I'm torn on whether putting it at a high or low level makes sense. I think low
level is better since it's otherwise duplicated in every operation. The
RECEIVED_SHUTDOWN signal also originals from near the record layer, when we see
an alert, so it makes some sense to check it there.  SENT_SHUTDOWN, in
contrast, is set at a fairly high-level "did I shut down the whole thing"
place.

Also, if we've seen a fatal alert, replay an error, not an EOF.

Change-Id: Idec35c5068ddabe5b1a9145016d8f945da2421cf