Update patch set 4
Patch Set 4:

> There's a jump in ERR_SSL_PROTOCOL_ERROR in Chrome canaries around
 > the vicinity of this change. (I'd normally suspect my SPKI
 > tightening, but that actually all went in before the jump.) I'm
 > going to revert this to see if it comes back down. If not, I'll
 > reland.

You might be in luck, Brian. :-) I've only gotten one data point back so far, but it doesn't look like it's this CL. I'll reland this if it holds for a couple more data points and search for another culprit.

(I swear, if it's one of the asm changes...)

Patch-set: 4