Update patch set 2
Patch Set 2:

I'll probably have both of you review this patch chain. Not because the CLs are especially large or complex, but because ssl3_read_bytes is a pretty core function in the SSL stack and it's probably worth using this patchset as some context.

Probably we should grab a VC and chat about that function and how TLS is structured. (I dunno how familiar you two are with it.)

Patch-set: 2