Update patch set 3
Patch Set 3:

> I don't have time to make the changes you suggest. Feel free to
 > adjust the commit as necessary or have somebody else do so.
 > (Sorry.)
 > 
 > In particular, I think it's wrong to just initialize a variable to
 > NULL to avoid the uninitialized value warnings without
 > understanding the code, and I have no desire whatsoever to try to
 > understand any of this code.

Fair enough. :-) I'll add it to my queue then. I'd normally agree with you on just initializing variables, but this is in the legacy pile of OpenSSL ASN.1 code which, like you, I have little interest in trying to understand carefully. So aligning with upstream (which just set them to zero) makes the most sense.

Also, from looking at the cases locally, a lot of them I think it would be correct to initialize the variable. Notably, if there's a pointer/length pair where the pointer is initialized to NULL, we should be initializing the length too.

Patch-set: 3
Label: Code-Review=0