Be more conservative about stack usage in |BN_mod_exp_mont_consttime|.
Allocate a single buffer that is used for all code paths in
|BN_mod_exp_mont_consttime|. Previously, the compiler might have been
allocating |powerbuf_stack| on the stack before calling the RSAZ
functions, instead of only allocating the space iif the RSAZ functions
aren't called. (Technically, it might have done so even with the
previous |alloca|-based implementation, but it is more likely to do it
now.) The RSAZ functions also use a large amount of stack space, and
this could result in twice as much stack space being used.

There are four main side benefits of this change:

* Less casting; in particular, fewer effective type rule violations.
* A larger stack-based buffer for the non-RSAZ implementation.
* A consistent amount of stack space being used (probably).
* Improved clarity of the code.

Change-Id: I1004795b2ad73da6e05a718260e99276a69cdba6