Drop support custom |mod_exp| hook in |RSA_METHOD|.
The documentation in |RSA_METHOD| says that the |ctx| parameter to
|mod_exp| can be NULL, however the default implementation doesn't
handle that case. That wouldn't matter since internally it is always
called with a non-NULL |ctx| and it is static, but an external
application could get a pointer to |mod_exp| by extracting it from
the default |RSA_METHOD|.

More importantly, this makes future refactorings easier.

Change-Id: Ie0e35e9f107551a16b49c1eb91d0d3386604e594