Update patch set 6
Patch Set 6:

Pfft. Apparently Conscrypt's current logic lets mix-and-match a lot of missing fields.
https://android.googlesource.com/platform/external/conscrypt/+/master/src/main/java/org/conscrypt/OpenSSLRSAPrivateCrtKey.java#81

So the thing that this would break currently works and is exposed as public API on Android. RSAPrivateKeyCrtKeySpec does not enforce that the components are not nullable.

Since it's actually less code to just fall back to the non-CRT codepath (which, thanks to the n/d case, will have to stick around regardless), I don't think it's really worthwhile to poke this compatibility bear. The Conscrypt maintainer also asked that we keep it working unless we have a good reason not to.

Will update the patch accordingly and land it.

Patch-set: 6
Label: Code-Review=0