Update patch set 6
Patch Set 6:

> > We should probably write code in Conscrypt to recover things
 > based on guesses of public exponents, but I think that sort of
 > thing should be either external or, at least, off to the side and
 > not in-line with the actual operations.
 > 
 > OK. In that case, though, I recommend:
 > 
 > (1) Provide the API for recovering `e` in BoringSSL, to minimize
 > direct `BIGNUM` API access outside BoringSSL.
 > 
 > (2) Once conscrypt has been changed to do that recovery of `e`,
 > your modifications to my other patch should be undone, so that the
 > operation fails if `e` is missing and the CRT parameters are
 > provided.

The other patch was also a little odd in that we're still tolerant of half the CRT parameters being missing, etc. In so far as we're okay with e-less keys hitting the non-CRT codepath at all, I think it seems reasonable to say that adding other components shouldn't break the key. (Unless there's some reason I'm missing that losing this case alone is valuable?)

There are other checks on e though that I do agree it'd be nice to lose lose. I would probably say modify (2) to say RSA private keys require n, e, and d and do a check for them early. (This would require that Conscrypt simply reject keys where e cannot be recovered, but I think that's probably reasonable? Haven't checked with Kenny on that.) At that point, we can actually cut down on codepaths. (Codepaths that disable security features no less, so it's pretty valuable.)

If we feel really adventurous, we can maybe try to recover p, q, and CRT parameters (is that possible?) and lose another codepath, but that's probably more trouble than is worth it since we're just getting an optimization back?

Patch-set: 6