Update patch set 6
Patch Set 6:

> The other patch was also a little odd in that we're still tolerant of half the CRT parameters being missing, etc.

Yes, that's weird and bad.

> In so far as we're okay with e-less keys hitting the non-CRT codepath at all, I think it seems reasonable to say that adding other components shouldn't break the key. (Unless there's some reason I'm missing that losing this case alone is valuable?)

Without `e`, you cannot verify the result is correct and you cannot use blinding, which are safety issues. Not being able to do the CRT optimization is not so much a safety issue.

> I would probably say modify (2) to say RSA private keys require n, e, and d and do a check for them early.

I agree, except that you don't need `d` at all if you're using the CRT.

> (This would require that Conscrypt simply reject keys where e cannot be recovered, but I think that's probably reasonable? Haven't checked with Kenny on that.)

It depends on how side-channel-resistant and complex the recovery operation is, and whether you need to support keys with uncommon values of `e`.

> If we feel really adventurous, we can maybe try to recover p, q, and CRT parameters (is that possible?) and lose another codepath

Yes. See, for example, https://github.com/pyca/cryptography/issues/975.

The issue comes down to offline vs online recovery of parameters: If the system can be configured so that the recovery operation only happens once per key (e.g. when the key is imported) then it definitely makes sense to just do recovery and CRT all the time. If, however, the recovery operation has to happen in line with the signing operation then it may add more risk than it alleviates due to side-channel concerns.

Patch-set: 6